requirements.txt
#pandas==3.0.0
python-dotenv==1.2.1
pydantic==2.12.5
tenacity==9.1.2
openai==2.26.0
google-genai==1.66.0

# Langchain Core
langchain==1.2.12
langchain-core==1.2.19

# Langchain Provider Integrations
langchain-openai==1.1.11
langchain-google-genai==4.2.1
langchain-community==0.4.1

# LangGraph
langgraph==1.1.2

# RAG & Vector Databases
chromadb==1.5.5
langchain-chroma==1.1.0
sentence-transformers==5.3.0

# UI
streamlit==1.55.0
pandas==2.3.3